[models/redis/user] + name refactor
