+++ app/model/__init__.py
@@ -1,2 +1,2 @@
-from .redis.redis_user import RedisUser
+from .redis.user import RedisUser
 from .constants import service_constants
